Annotate TLS related values in values.yaml
Comments have been added to the TLS section of the `values.yaml` file to improve understanding. These annotations address enabling TLS transport, the process of auto-generating self-signed certificates, the procedure for naming secrets that contain certificates, and the methods for defining paths and content for certificate files when mounted as secrets.

Signed-off-by: Jiří Altman <jiri.altman@usu.com>