New template dependencytrack.tpl.render
A new template for rendering Helm templates in values has been added to the _helpers.tpl file. This function takes a value that may contain a Helm template and evaluate it, which is useful for dynamic configuration values that need to be evaluated during chart rendering. The function also handles different data types and scopes, providing flexibility in its usage. An alias "!TPL" has been created for this function as well.

Signed-off-by: Jiří Altman <jiri.altman@usu.com>