Add support for extra Kubernetes resources
Introduced a new feature that allows users to deploy additional Kubernetes resources not included in the main chart package. This extends the capabilities of the chart by enabling custom resource specification. The template iterates over each item in `$.Values.extraDeploy`, passing it through a rendering process, and checks for valid YAML syntax. If all checks pass, the item is rendered as YAML multi-document and printed to output.

Signed-off-by: Jiří Altman <jiri.altman@usu.com>